Create separate data base file, parallelized webscraper, and added instructions for fillings up database with class info for the first time.
