+++ Instructions.txt
@@ -4,3 +4,8 @@ I finally did it and it was painful. To start run the command `docker-compose up
 You will see lots of red errors but trust the process. Once it is done you should be able to view the containers from localhost:5000 for backend
 and localhost:3000 for frontend. I set up some code to test it out and you should see that the message from the backend is sent to the front end
 via an api call. If it works then you did it and now you have a working docker enviroment.
+
+You can fill up the data base up by setting FILL_DB_WITH_CLASS_DATA to True in app.py. The webscraper is parallelized, so the semesters will be out of order.
+It still takes a while to fill up, but it is done once you see everything in the database (you can use the SQLite Viewer extension).
+Also, since it is parallelized, the server is slower while webscraping, which is not ideal, but I will keep it like that for now since I want it
+to get done as quick as possible. You only need to do this once since the database is persistent.
